Release-0.9.1 New Folder Docs
All the materials of the latest release version of the CoDesignTool, with changes on the instructions and boards/canvases. These adjustments are based on insights gathered from the testing session held on April 24th at the University of Bologna, aiming to improve user experience and effectiveness in the codesign process.